+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+0.0.2 / unreleased
+==================
+
+* Rename `Starlark.read` to `Starlark.exec`
+* Fix memory leak from not freeing C.CString
+
 0.0.1 / 2020-11-22
 ==================